更改v2为float32类型

+++ utils/maths.go
@@ -19,8 +19,8 @@ func MinInt32(i, j int32) int32 {
 	return j
 }
 
-func FloatEqual(i, j float64) bool {
-	return math.Abs(i-j) < 10e-8
+func FloatEqual(i, j float32) bool {
+	return math.Abs(float64(i-j)) < 10e-8
 }
 
 //RandomFloat64 随机生成一个min到max的float64的数
